requirements.txt
boto3
python-dotenv
oracledb
oci
pytest
moto[s3]
